requirements.txt
advertorch==0.2.3
advertorch.egg==info
matplotlib==3.5.2
numpy==1.21.5
pandas==1.3.5
Pillow==9.4.0
scipy==1.7.3
seaborn==0.12.2
skimage==0.0
torch==1.12.1
torchvision==0.13.1
wand==0.6.11
